+++ src/widgets/OverlapWidget.tsx
@@ -6,9 +6,13 @@ import type { CommonInternalProps } from './utils/common-internal.props';
 import type { CommonStyleProps } from './utils/style.props';
 import { convertCommonStyle } from './utils/style.utils';
 
-type OverlapWidgetInternalProps = CommonInternalProps;
+interface OverlapWidgetInternalProps extends CommonInternalProps {
+  overflow?: 'hidden';
+}
 
-export type OverlapWidgetStyle = CommonStyleProps;
+export interface OverlapWidgetStyle extends CommonStyleProps {
+  overflow?: 'hidden';
+}
 
 export interface OverlapWidgetProps extends ClickActionProps {
   children?: any;
@@ -25,5 +29,6 @@ OverlapWidget.convertProps = (
   return {
     ...convertCommonStyle(props.style ?? {}),
     ...convertClickAction(props),
+    ...(props?.style?.overflow ? { overflow: props.style.overflow } : {}),
   };
 };